integration: move skip_metadata_optimization to prepared
The test is strictly about prepared statements, so it's moved to the
corresponding module.
